requirements.txt
alembic==1.8.1
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
fastapi==0.100
uvicorn==0.18.3
google-cloud-logging==3.2.5
google-cloud-error-reporting==1.9.1
google-api-python-client==2.66.0
google-cloud-bigquery==3.4.0
google-cloud-storage==2.7.0
pytz==2022.7
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-exporter-gcp-trace==1.4.0
opentelemetry-exporter-gcp-monitoring==1.5.0a0
pydantic==1.10.8
psycopg2==2.9.6
cloud-sql-python-connector[pg8000]==1.2.3
google-cloud-iam==2.12.1
google-cloud-resource-manager==1.10.2
pytest